Add files via upload
Project submission for the portfolio project for freecodecamp.org. Includes links to other projects that are inactive in this repository. The contact links also do not lead anywhere. The reasons for this are that this is a 1-1 upload to serve as a record of my project on GitHub from when it was created months ago.